Add ability for multi-room polls
This changes how the data is persisted. Instead of have an "activePoll" pointer that points to the active poll, the key for the poll save is now activePoll_[channelId]. Channel_id comes from the slack post data. This allows us to have multiple polls running. There is only one poll allowed per slack channel. When using the /start route a new poll will be started for the specified channel. If one already exists for the channel, it will print the results of the existing poll, close it and start the new poll

Signed-off-by: Hans Gutknecht <hansdotcom@gmail.com>